Additional commit for about BoMs. Consolidate about pages
