+++ modules/page_object.py
@@ -1,10 +1,4 @@
-from modules.page_object_about_config import *
-from modules.page_object_about_downloads import *
-from modules.page_object_about_glean import *
-from modules.page_object_about_logins import *
-from modules.page_object_about_privatebrowsing import *
-from modules.page_object_about_profiles import *
-from modules.page_object_about_telemetry import *
+from modules.page_object_about_pages import *
 from modules.page_object_addons_mozilla_org import *
 from modules.page_object_autofill_credit_card import *
 from modules.page_object_autofill_login import *
