--- modules/page_object_about_privatebrowsing.py
@@ -1,5 +1,0 @@
-from modules.page_base import BasePage
-
-
-class AboutPrivatebrowsing(BasePage):
-    URL_TEMPLATE = "about:privatebrowsing"
