--- modules/page_object_about_profiles.py
@@ -1,9 +1,0 @@
-from modules.page_base import BasePage
-
-
-class AboutProfiles(BasePage):
-    """
-    POM for about:profiles page
-    """
-
-    URL_TEMPLATE = "about:profiles"
